Refactor AgeRangeSubjectsForm
This extracts the logic of the age range subjects form in to a concern
which we can use in multiple places.
